+++ include/events/EventHandler.h
@@ -7,6 +7,9 @@
 #include <QObject>
 #include <QSharedPointer>
 #include <QScopedPointer>
+#include <QLoggingCategory>
+
+Q_DECLARE_LOGGING_CATEGORY(event_hyperion);
 
 class Logger;
 
@@ -19,6 +22,9 @@ public:
 
 	static QScopedPointer<EventHandler>& getInstance();
 	static void destroyInstance();
+
+	bool isSuspended() { return _isSuspended; }
+	bool isIdle() { return _isIdle; }
 
 public slots: