+++ include/grabber/amlogic/AmlogicGrabber.h
@@ -45,7 +45,7 @@ public:
 	///
 	/// @return A JSON structure holding a list of devices found
 	///
-	QJsonObject discover(const QJsonObject &params);
+	QJsonObject discover(const QJsonObject &params) const;
 
 	///
 	/// Set the video mode (2D/3D)
@@ -95,6 +95,8 @@ private:
 
 	Image<ColorBgr> _image_bgr;
 	ColorBgr* _image_ptr;
+	size_t _stride;
+	size_t _bytesToRead;
 
 	int _lastError;
 	bool _videoPlaying;