+++ libsrc/events/EventHandler.cpp
@@ -7,6 +7,8 @@
 #include <utils/Process.h>
 #include <hyperion/HyperionIManager.h>
 
+Q_LOGGING_CATEGORY(event_hyperion, "hyperion.event.hyperion");
+
 QScopedPointer<EventHandler> EventHandler::instance;
 
 EventHandler::EventHandler()
@@ -72,12 +74,11 @@ void EventHandler::suspend(bool sleep)
 	}
 	else
 	{
-		if (_isSuspended || _isIdle)
+		if (_isSuspended)
 		{
 			Info(_log, "Resume event received - Hyperion is going into working mode");
-			emit signalEvent(Event::Resume);
 			_isSuspended = false;
-			_isIdle = false;
+			emit signalEvent(Event::Resume);
 		}
 		else
 		{
@@ -127,8 +128,8 @@ void EventHandler::idle(bool isIdle)
 			if (_isIdle)
 			{
 				Info(_log, "Resume from idle event recevied");
-				emit signalEvent(Event::ResumeIdle);
 				_isIdle = false;
+				emit signalEvent(Event::ResumeIdle);
 			}
 		}
 	}
@@ -205,6 +206,14 @@ void EventHandler::handleEvent(Event event)
 
 	case Event::Quit:
 		emit signalEvent(Event::Quit);
+		break;
+
+	case Event::Lock:
+		emit signalEvent(Event::Lock);
+		break;
+
+	case Event::Unlock:
+		emit signalEvent(Event::Unlock);
 		break;
 
 	default: